requirements.txt
numpy
scipy
pandas
loguru
pillow
torchseg
matplotlib
scikit-image
opencv-python